genai-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
langchain==0.3.25
langchain-openai==0.3.23
langgraph==0.4.8
python-dotenv==1.1.0
pydantic==2.11.7